+++ packages/labnative/pages/Result.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { SearchList, Avatar } from '@gympass/yoga';
+import { Result, Avatar } from '@gympass/yoga';
 import { BuildingFilled } from '@gympass/yoga-icons';
 import { DocTitle } from '../components';
 
@@ -22,7 +22,7 @@ const ScrollView = styled.ScrollView`
   width: 100%;
 `;
 
-const SearchListWrapper = styled.View`
+const ResultWrapper = styled.View`
   align-items: center;
   justify-content: space-between;
   flex-direction: column;
@@ -36,30 +36,19 @@ const SearchListWrapper = styled.View`
 
 const ResultPage = () => (
   <ScrollView>
-    <SearchListWrapper>
-      <DocTitle>Default Avatar without image prop</DocTitle>
-      <SearchList>
-        <SearchList.Header
-          rate="5.0"
-          avatar={<Avatar />}
-          attendanceList={attendanceList}
-        >
-          Partner Name
-        </SearchList.Header>
-      </SearchList>
-    </SearchListWrapper>
-    <SearchListWrapper>
-      <DocTitle>Default Avatar without image prop</DocTitle>
-      <SearchList>
-        <SearchList.Header
-          rate="5.0"
-          avatar={<Avatar.Circle />}
-          attendanceList={attendanceList}
-        >
-          Partner Name
-        </SearchList.Header>
-      </SearchList>
-    </SearchListWrapper>
+    <ResultWrapper>
+      <DocTitle>Default Result</DocTitle>
+      <Result
+        rate="5.0"
+        avatar={<Avatar.Circle />}
+        attendances={attendanceList}
+        title="Partner Name"
+        subTitle="Activity"
+        divided
+      >
+        <Result.Button>Link Button</Result.Button>
+      </Result>
+    </ResultWrapper>
   </ScrollView>
 );
