--- packages/yoga/src/Result/web/index.js
@@ -1,3 +1,0 @@
-import Result from './SearchList';
-
-export default Result;
